Move database fixture into conftest so it can be used elsewhere.
